WMS: implement deferred driver loading
