+++ frmts/wms/wmsdriver.cpp
@@ -43,6 +43,7 @@
 #include "minidriver_mrf.h"
 #include "minidriver_ogcapimaps.h"
 #include "minidriver_ogcapicoverage.h"
+#include "wmsdrivercore.h"
 
 #include "cpl_json.h"
 
@@ -716,81 +717,6 @@ static CPLXMLNode *GDALWMSDatasetGetConfigFromArcGISJSON(const char *pszURL,
 }
 
 /************************************************************************/
-/*                             Identify()                               */
-/************************************************************************/
-
-int GDALWMSDataset::Identify(GDALOpenInfo *poOpenInfo)
-{
-    const char *pszFilename = poOpenInfo->pszFilename;
-    const char *pabyHeader = (const char *)poOpenInfo->pabyHeader;
-    if (poOpenInfo->nHeaderBytes == 0 &&
-        STARTS_WITH_CI(pszFilename, "<GDAL_WMS>"))
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes >= 10 &&
-             STARTS_WITH_CI(pabyHeader, "<GDAL_WMS>"))
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes == 0 &&
-             (STARTS_WITH_CI(pszFilename, "WMS:") ||
-              CPLString(pszFilename).ifind("SERVICE=WMS") != std::string::npos))
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes != 0 &&
-             (strstr(pabyHeader, "<WMT_MS_Capabilities") != nullptr ||
-              strstr(pabyHeader, "<WMS_Capabilities") != nullptr ||
-              strstr(pabyHeader, "<!DOCTYPE WMT_MS_Capabilities") != nullptr))
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes != 0 &&
-             strstr(pabyHeader, "<WMS_Tile_Service") != nullptr)
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes != 0 &&
-             strstr(pabyHeader, "<TileMap version=\"1.0.0\"") != nullptr)
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes != 0 &&
-             strstr(pabyHeader, "<Services") != nullptr &&
-             strstr(pabyHeader, "<TileMapService version=\"1.0") != nullptr)
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes != 0 &&
-             strstr(pabyHeader, "<TileMapService version=\"1.0.0\"") != nullptr)
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes == 0 &&
-             STARTS_WITH_CI(pszFilename, "http") &&
-             (strstr(pszFilename, "/MapServer?f=json") != nullptr ||
-              strstr(pszFilename, "/MapServer/?f=json") != nullptr ||
-              strstr(pszFilename, "/ImageServer?f=json") != nullptr ||
-              strstr(pszFilename, "/ImageServer/?f=json") != nullptr))
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes == 0 &&
-             STARTS_WITH_CI(pszFilename, "AGS:"))
-    {
-        return TRUE;
-    }
-    else if (poOpenInfo->nHeaderBytes == 0 &&
-             STARTS_WITH_CI(pszFilename, "IIP:"))
-    {
-        return TRUE;
-    }
-    else
-        return FALSE;
-}
-
-/************************************************************************/
 /*                                 Open()                               */
 /************************************************************************/
 
@@ -802,7 +728,7 @@ GDALDataset *GDALWMSDataset::Open(GDALOpenInfo *poOpenInfo)
     const char *pszFilename = poOpenInfo->pszFilename;
     const char *pabyHeader = (const char *)poOpenInfo->pabyHeader;
 
-    if (!Identify(poOpenInfo))
+    if (!WMSDriverIdentify(poOpenInfo))
         return nullptr;
 
     if (poOpenInfo->nHeaderBytes == 0 &&
@@ -1133,65 +1059,6 @@ void WMSDeregister(CPL_UNUSED GDALDriver *d)
     WMSRegisterMiniDriverFactory(new WMSMiniDriverFactory_##name());
 
 /************************************************************************/
-/*                    OGRWMSDriverGetSubdatasetInfo()                   */
-/************************************************************************/
-
-struct WMSDriverSubdatasetInfo : public GDALSubdatasetInfo
-{
-  public:
-    explicit WMSDriverSubdatasetInfo(const std::string &fileName)
-        : GDALSubdatasetInfo(fileName)
-    {
-    }
-
-    // GDALSubdatasetInfo interface
-  private:
-    void parseFileName() override
-    {
-        if (!STARTS_WITH_CI(m_fileName.c_str(), "WMS:"))
-        {
-            return;
-        }
-
-        const CPLString osLayers = CPLURLGetValue(m_fileName.c_str(), "LAYERS");
-
-        if (!osLayers.empty())
-        {
-            m_subdatasetComponent = "LAYERS=" + osLayers;
-            m_driverPrefixComponent = "WMS";
-
-            m_pathComponent = m_fileName;
-            m_pathComponent.erase(m_pathComponent.find(m_subdatasetComponent),
-                                  m_subdatasetComponent.length());
-            m_pathComponent.erase(0, 4);
-            const std::size_t nDoubleAndPos = m_pathComponent.find("&&");
-            if (nDoubleAndPos != std::string::npos)
-            {
-                m_pathComponent.erase(nDoubleAndPos, 1);
-            }
-            // Reconstruct URL with LAYERS at the end or ModifyPathComponent will fail
-            m_fileName = m_driverPrefixComponent + ":" + m_pathComponent + "&" +
-                         m_subdatasetComponent;
-        }
-    }
-};
-
-static GDALSubdatasetInfo *WMSDriverGetSubdatasetInfo(const char *pszFileName)
-{
-    if (STARTS_WITH(pszFileName, "WMS:"))
-    {
-        std::unique_ptr<GDALSubdatasetInfo> info =
-            std::make_unique<WMSDriverSubdatasetInfo>(pszFileName);
-        if (!info->GetSubdatasetComponent().empty() &&
-            !info->GetPathComponent().empty())
-        {
-            return info.release();
-        }
-    }
-    return nullptr;
-}
-
-/************************************************************************/
 /*                          GDALRegister_WMS()                          */
 /************************************************************************/
 
@@ -1204,7 +1071,7 @@ static GDALSubdatasetInfo *WMSDriverGetSubdatasetInfo(const char *pszFileName)
 void GDALRegister_WMS()
 
 {
-    if (GDALGetDriverByName("WMS") != nullptr)
+    if (GDALGetDriverByName(DRIVER_NAME) != nullptr)
         return;
 
     // Register all minidrivers here
@@ -1221,19 +1088,11 @@ void GDALRegister_WMS()
     RegisterMinidriver(OGCAPICoverage);
 
     GDALDriver *poDriver = new GDALDriver();
-
-    poDriver->SetDescription("WMS");
-    poDriver->SetMetadataItem(GDAL_DCAP_RASTER, "YES");
-    poDriver->SetMetadataItem(GDAL_DMD_LONGNAME, "OGC Web Map Service");
-    poDriver->SetMetadataItem(GDAL_DMD_HELPTOPIC, "drivers/raster/wms.html");
-    poDriver->SetMetadataItem(GDAL_DCAP_VIRTUALIO, "YES");
-    poDriver->SetMetadataItem(GDAL_DMD_SUBDATASETS, "YES");
+    WMSDriverSetCommonMetadata(poDriver);
 
     poDriver->pfnOpen = GDALWMSDataset::Open;
-    poDriver->pfnIdentify = GDALWMSDataset::Identify;
     poDriver->pfnUnloadDriver = WMSDeregister;
     poDriver->pfnCreateCopy = GDALWMSDataset::CreateCopy;
-    poDriver->pfnGetSubdatasetInfoFunc = WMSDriverGetSubdatasetInfo;
 
     GetGDALDriverManager()->RegisterDriver(poDriver);
 }
